It all comes from focusing, I tell you
When trying to integrate sum and product in type descriptors, an
asymmetry between positive and negative types arose -- while we
previously only supported the latter. I'm not sure yet whether it is
fundamental, or we may restore a symmetrical situation by a different
point of view on function signatures.

Note that actual examples are not yet converted to this richer type
language.