move the Obj.magic to a different place by using a laaarge functor
The idea is to isolate the crucial assumption that Thomas was making
in his previous use of Obj.magic: if two type descriptions have the
same unique identifier, then they describe the exact same type. He was
using this meta-result with an Obj.magic right in the middle of
a dynamic graph-based imperative weak-hash-table gc-clever algorithm,
which is a bit too smart to be reasonable. The functor moves this
assumption out of this part of the code, to be asserted (still with
Obj.magic) separately.

The good thing is that we can now think of different ways to prove
this assumption, without Obj.magic: for any given signature, we can in
fact declare a home-grown ('a dynty) GADT that would let the
type-checker realize that the assumption is true. This is the goal for
the next commit(s).